move StaminaHUD class package
