+++ src/main/java/com/alrex/parcool/client/hud/AbstractHUD.java
@@ -4,7 +4,7 @@ import com.mojang.blaze3d.matrix.MatrixStack;
 import net.minecraft.client.gui.AbstractGui;
 
 public abstract class AbstractHUD extends AbstractGui {
-	AbstractHUD(Position pos) {
+	public AbstractHUD(Position pos) {
 		position = pos;
 	}
